fix: parsing bio (#360)
